healthserver/Backend_health_assistant/requirements.txt
Flask==3.0.3
google-generativeai==0.5.0
pymongo==4.8.0
gtts==2.5.1
python-dotenv==1.0.1
flask-cors==4.0.1